Add movement reducer to shipment reducer and move movement reducers into its own folder
