+++ imports/state/reducers/shipment/shipmentReducer.js
@@ -1,5 +1,7 @@
 import { cargo } from '../cargo/cargoReducers';
+import { movement } from '../movement/movementReducers';
 
 export const shipment = (state = {}, action) => ({
   cargo: cargo(state.cargo || {}, action),
+  movement: movement(state.movement || {}, action),
 });
